Improve authorization enforcement in fir_artifacts views

+++ fir_artifacts/views.py
@@ -1,15 +1,16 @@
 from django.contrib.auth.decorators import login_required, user_passes_test
+from django.core.exceptions import PermissionDenied
 from django.http import Http404
 from django.shortcuts import get_object_or_404, render, redirect
 from fir_artifacts.files import do_download_archive, do_download, do_upload_file, do_remove_file
 
-from incidents.views import is_incident_handler
+from incidents.views import is_incident_viewer
 
 from fir_artifacts.models import Artifact
 
 
 @login_required
-#@user_passes_test(is_incident_handler)
+@user_passes_test(is_incident_viewer)
 def artifacts_correlations(request, artifact_id):
     a = get_object_or_404(Artifact, pk=artifact_id)
     correlations = a.relations.group()
@@ -17,7 +18,6 @@ def artifacts_correlations(request, artifact_id):
 
 
 @login_required
-@user_passes_test(is_incident_handler)
 def detach_artifact(request, artifact_id, relation_name, relation_id):
     a = get_object_or_404(Artifact, pk=artifact_id)
     relation = getattr(a, relation_name, None)
@@ -27,6 +27,8 @@ def detach_artifact(request, artifact_id, relation_name, relation_id):
         related = relation.get(pk=relation_id)
     except:
         raise Http404("Unknown related object")
+    if not request.user.has_perm('incidents.handle_incidents', obj=related):
+        raise PermissionDenied()
     a.relations.remove(related)
     if a.relations.count() == 0:
         a.delete()
@@ -34,24 +36,20 @@ def detach_artifact(request, artifact_id, relation_name, relation_id):
 
 
 @login_required
-@user_passes_test(is_incident_handler)
 def download_archive(request, content_type, object_id):
     return do_download_archive(request, content_type, object_id)
 
 
 @login_required
-@user_passes_test(is_incident_handler)
 def download(request, file_id):
     return do_download(request, file_id)
 
 
 @login_required
-@user_passes_test(is_incident_handler)
 def upload_file(request, content_type, object_id):
     return do_upload_file(request, content_type, object_id)
 
 
 @login_required
-@user_passes_test(is_incident_handler)
 def remove_file(request, file_id):
     return do_remove_file(request, file_id)
